% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_data.R
\docType{data}
\name{BPUrban300}
\alias{BPUrban300}
\title{Hypertension, blood pressure and their risk factors in 300 participants}
\format{
A data frame with 150 rows and 23 variables:
\describe{
\item{hypertension}{is the participant hypertensive? 1 =  Yes, 0 = No}
\item{systolic}{systolic blood pressure (mmHg)}
\item{diastolic}{diastolic blood pressure (mmHg)}
\item{bmi}{body mass index}
\item{bmi_group}{BMI binned}
\item{cholesterol}{cholesterol (mg/dL)}
\item{age}{age of participant (years)}
\item{age_group}{age binned}
\item{parent_history_HTN}{parent history of hypertension}
\item{salt}{salt intake}
\item{salt_group}{salt binned}
\item{smoking}{does the participant smoke cigarettes?}
\item{chewing}{does the participant chew tobacco?}
\item{activity}{physical activity}
\item{weight}{weight}
\item{waist_circumference}{waist circumference}
\item{hip_circumference}{hip circumference}
\item{waist_hip_ratio}{waist to hip ratio}
\item{fasting_glucose}{fasting glucose level (mmol/L)}
\item{caste}{caste}
\item{marital_status}{marital status}
\item{religion}{religion}
\item{education}{education}
}
}
\usage{
BPUrban300
}
\description{
A simulated data set containing hypertension, blood pressure measurements, and
potential risk factors for these two outcomes.
}
\keyword{datasets}
