% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_data.R
\docType{data}
\name{BPUrban500}
\alias{BPUrban500}
\title{Hypertension, blood pressure and their risk factors in 500 participants}
\format{
A data frame with:
\describe{
\item{hypertension}{Hypertensive status. Is the participant hypertensive? 1 = Yes, 0 = No}
\item{systolic}{Systolic blood pressure (mmHg)}
\item{bmi}{Body mass index (kg/m^2)}
\item{age}{Age (years)}
\item{age_g66}{Age greater than mean = 66 years}
\item{smoking}{Does he/she smoke? T/F}
}
}
\usage{
BPUrban500
}
\description{
A simulated data set containing hypertension, blood pressure measurements, and
potential risk factors of these two outcomes.
}
\keyword{datasets}
