% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_data.R
\docType{data}
\name{growth_data}
\alias{growth_data}
\title{Age, weight, and height of two newborns}
\format{
A data frame with 8 rows and 4 variables:
\describe{
\item{id}{three-number subject ID}
\item{months}{age in months at visit}
\item{weight_kg}{person's weight in kilograms}
\item{height_cm}{person's height in centimeters}
}
}
\usage{
growth_data
}
\description{
A made-up dataset containing an id, age, weight, and height of two newborns
followed monthly for the first three months after birth. These data are
used during the videos to demonstrate data wrangling ideas.
}
\keyword{datasets}
