% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_data.R
\docType{data}
\name{growth_data}
\alias{growth_data}
\title{Age, weight, and height of 8 made-up people}
\format{
A data frame with 8 rows and 3 variables:
\describe{
\item{age_group}{person's age category}
\item{weight_kg}{person's weight in kilograms}
\item{height_cm}{person's height in centimeters}
}
}
\usage{
growth_data
}
\description{
A made-up dataset containing the age, weight, and height of 8 people. It is
used during the videos to demonstrate data wrangling ideas.
}
\keyword{datasets}
